+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The present file will list all changes made to the project; according to the
 - Alerts on saved searches
 - Add ajax browsers notifications in addition to emails
 - Plugins can now add new notifications types (xmpp, sms, telegram, ...) to be used along with standard notifications
+- Simcard component
 
 ### Changed
 - Display knowledge base category items in tickets using a popup instead of a